requirements.txt
streamlit
pillow
python-dotenv
requests
langchain-huggingface
diffusers
torch
accelerate
safetensors
transformers 